test_task/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
drf-excel==2.1.0
et-xmlfile==1.1.0
greenlet==1.1.2
humanize==4.1.0
idna==3.3
kombu==5.2.4
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
